backend/requirements.txt
Django>=4.0

djangorestframework
djangorestframework-simplejwt

django-cors-headers

PyJWT==2.8.0
python-dotenv==1.0.0

google-generativeai==0.3.2

psycopg2-binary==2.9.9
requests==2.31.0
